move undo scope to its own file
